CA-310971: consider enabling the host again after plugging clustering PBDs
When HA is enabled the host only gets enabled if all its PBDs got
plugged.
However clustered PBDs only get plugged after quorum is established,
which might be after the startup sequence has finished.

There is a watcher thread for that, and when the clustered PBDs do get
plugged in we should attempt to enable the host again.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>