CP-30440: always set pool certificate
While developing the certificate upload code it is convenient to have the latest
host certificate used by the pool.
We have no way of wiping the pool certificate, so this is probably the right approach
in general: if we ever fix a bug in the certificate parsing upload code
we need a way to propagate it to the pool.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>