CP-30440: fix tarfile extraction
We need to write out the tarfile before extracing it,
otherwise there is nothing to extract.
Use the actual filename inside the tarfile, otherwise we try to extract
a file into a directory which fails with EISDIR.

TODO: we will need to sanitize the filename here and check it has no dotdot in it.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>