CA-311823: Do not raise alerts for "new CPU features"
These alerts may cause confusion, because sometimes the "new features" are not
actually new hardware features, but ones that are newly exposed by Xen to
guests.

We are keeping the "lost features" alerts, which are far more useful, and also
extend the logging to include the old as well as the new features.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>