CA-310525 fix C binding for statvfs
We rarely see statvfs return 0 for bfree and suspect that this is a GC
related problem. The existing implementation violates rule 6 in the
OCaml manual

  https://caml.inria.fr/pub/docs/manual-ocaml/intfc.html#sec442

which mandates to use caml_modify when updating a field with an existing
value. This commit handles this by using Store_field. See also the
discussion at

  https://discuss.ocaml.org/t/finding-gc-problems-in-c-bindings/3453/2

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>