CA-313709: Save xen-platform QEMU params in persistent VM state
This means that these params do not get lost across migrations and
suspend/resume cycles, as the VM state is kept with the VM.

To do this, the params are derived at a higher level in `create_device_model`
in Xenops_server_xen, rather than down in the code that executes QEMU
(Device.Dm). This means that some plumbing is needed, but so far there is no
change in behaviour.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>