CA-314001: release runtime lock around long running system calls
In CA-314001 VM import was running fsync, which caused all other threads
in XAPI to be blocked, which in turn caused the HA xapi health checker
to fence the host.

We should release the runtime lock whenever we perform a syscall,
especially those that perform I/O, since they can block for a long time
(minutes) if there are a lot of dirty pages and the storage is slow.

The change here follows the recommendations from the manual
section 20.12.2 Parallel execution of long-running C code
http://caml.inria.fr/pub/docs/manual-ocaml/intfc.html

Signed-off-by: Edwin Török <edvin.torok@citrix.com>