CA-314170: Revert "Don't wait for hotplug-status to disappear for qdisk"
This reverts commit cbc66c70051de63d8221b5cd6d86cb2c9f0f2245.

The dummy unplug watch "fired" immediately, causing xenopsd to delete
to delete the xenstore keys before the frontend had closed it.
It also caused xenopsd to not detect the `error` from xenstore that the
device is in use and cannot be unplug.
This caused both the backend and frontend to get confused about the
state of the device and wrote/processed random garbage from the PV ring.

If the domain is still alive (i.e. clean shutdown of a device) then we
need to wait for the frontent to reach the Closed state before
unplugging, which the code would do if unplug doesn't fire too soon.

Instead the udev rule should get triggerred for qemu-dp(qdisk) the same
way it is triggerred for tapdisk(vbd*).

Signed-off-by: Edwin Török <edvin.torok@citrix.com>