CA-312226 XSI-251 clear unexpected vGPU metadata on shutdown
A vGPU is supposed to hold vGPU compatibility metadata only while it is
suspended. Beyond that, it can deadlock the start of other machines as
it causes gpumon being contacted in vgpu_pgpu_are_compatible() while
gpumon might be is stopped as part of VM.start of another VM. This
commit adds code to detect, log, and remove this metadata when a VM is
halted to avoid the problem.  We still need to find the root cause why
the situation can arise in the first place.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>