CA-314511: Set "type" key for HVM domains that is expected by xl
Most of xl utilities while quering for domain type are trying to
distinguish between HVM and PVH by reading /libxl/<domid>/type key.
It now mostly works because the default behavior is to assume the
domain is HVM and show a warning.

Set the key explicitly to "HVM" to avoid the warning and confusion.

Signed-off-by: Igor Druzhinin <igor.druzhinin@citrix.com>