CA-314170: remove hotplug status when backend is closed on clean shutdown (qdisk)
There is no physical device in Dom0 for qemu-dp (qdisk) devices,
so we have to do this ourselves here.
When the frontend reaches state Closed, the backend will go to state Closed,
at which point it should be safe to delete the hotplug status key.
Deleting this key will cause the udev rules to fire, the same way as they would for
a tapdisk device (vbd3).

We need to evaluate all the watches, so use Watch.any_of:
* the device might already be unplugged in which case the unplug_watch
would fire
* the backend may not be closed yet, in which case once the backend
closed watch fires we will remove the hotplug status key
* Watch.map is only evaluated if the watch itself is evaluated,
so using Watch.all would timeout because it would short-circuit
evaluation on the first failing watch

Signed-off-by: Edwin Török <edvin.torok@citrix.com>