CA-314170: remove hotplug status immediately on hard shutdown (qdisk)
On a hard shutdown we cannot rely on the domain's cooperation,
so remove the hotplug status key immediately for qdisk backends.
This restores the behaviour reverted in 0037eefd038d3f65289cefcf36a38f859dc19763,
but only for hard shutdown, but it avoids changing the watch itself.
(Changing the watch itself affects clean shutdown code too, which causes
 CA-314170)

Signed-off-by: Edwin Török <edvin.torok@citrix.com>