CA-315107 Create xapi-init-complete systemd target
Create a xapi-init-complete.target and a xapi-wait-init-complete.service
which is wanted by it. The service invokes the existing
xapi-wait-init-complete so that it is only done in one place now.

The SPEC file will need altering to enable the xapi-wait-init-complete
service on installation.

xapi-domains.service (and any other service which needs to start after
Xapi has completed initialisation) can now declare

Wants=xapi-init-complete.target
After=xapi-init-complete.target

to ensure that Xapi will be done initialising before they are started.

Signed-off-by: Tim Smith <tim.smith@citrix.com>