CA-315952 Use Ezjsonm for json serialisation
This commit addresses problems with the json serialisation: The existing
serialisation is naive and does not respect escape rules for Json - for
example a string containing a backslash or quote character must be
escaped in its Json representation. We replace the naive serialisation
by building an abstract Json representation and pass it to Ezjsonm for
serialisation - which takes care of implementing these rules.

In addition, the interface to rrd_json is slimmed down by removing two
unused functions:

* val json_of_ds
* val json_metadata_of_ds

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>