CA-316241 redirect stderr output from probe-device-for-file
probe-device-for-file is called from xe-restore-metadata which is called
from xsconsole. Output to stderr, stdin is shown to the user and it
contains irrelevant and confusing output from probe-device-for-file
which is coming from the fsimage library. Since we can't change the
fsimage library, stderr for the call to fsimage (only) is redirected to
/dev/null. This does not make us miss errors as these are raised as
exceptions.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>