Remove the Xenctrl dependency in Monitor_master.update_pifs
The code that handles the `pass_through_pif_carrier` feature finds all VIFs
that are on any bridges that are associated with a PIF that has seen a change
in carrier. For all those VIFs, the `VIF.set_carrier` function of xenopsd is
called.

This is done by listing all the interfaces on those bridges and finding the VIF
devices by matching the names with `vif<domid>.<devid>`. Then, to call xenopsd,
we need to know the UUID of the VM to which the VIF belongs. Xapi used to ask
Xen, but now looks this up in its own database.

This patch also simplifies the code somewhat.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>