CA-315952 fix JSON output
Json so far is generated in a naive way that violates the Json
specification by ignoring basic quoting rules. This commit delegates
generation of Json to a library instead and should be much more
declarative.

However, the generated JSON still is not equivalent to the XML as it
contains hard-coded values in some places where the XML seems to hold
actual data. As can be seen from the code,

  * last_ds:0.0,
  * value:0.0,
  * unknown_sec:0

are hard coded in the Json output but not in the XML output.

   {
      "name": "memory_total_kib",
      "type": "GAUGE",
      "minimal_hearbeat": "300.0000",
      "min": "0.0",
      "max": "Infinity",
      "last_ds": "0.00",
      "value": "0.00",
      "unknown_sec": "0.00"
    },

  <ds>
    <name>memory_total_kib</name>
    <type>GAUGE</type>
    <minimal_heartbeat>300.0000</minimal_heartbeat>
    <min>0.0</min>
    <max>Infinity</max>
    <last_ds>33548544</last_ds>
    <value>160004678.0795</value>
    <unknown_sec>0</unknown_sec>
  </ds>

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>