CA-320361: use device-id if it exists
For compatibility with old tools and Naples.
device-id is still 2 in the template (and in xenstore) for UEFI,
but not in qemu, which confuses things.

The template can be updated to remove the device-id in the future,
however it cannot be done yet because the old tools fail with device-id
1.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>