CA-321983: Write compatibility lookup file for NVidia VGPU types
The file contains a mapping of config file paths to the type_id and UUID of the
VGPU type, which can be used by xenopsd to translate the former to the latter,
when receiving a VM from an older host. The old-style NVidia VGPU metadata in
xenopsd contains just the config file path, but nowadays xenopsd needs the
extra parameters.

This brings back some of the code to parse the NVidia config files, which was
recently deleted, but in a cut-down form.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>