CA-321983: Handle vGPU migration from older releases
The NVidia vGPU metadata that comes from older xenopsds contains a config_file
path rather than a type_id and VGPU UUID. Match config_file and type_id up in
the lookup file that xapi has helpfully created to handle this case. The VGPU
UUID is not available at all, but we just create a new UUID. XAPI uses the UUID
to distinguish between multiple VGPUs on a VM. However, there is only one VGPU
per VM in this upgrade case, so the UUID is not really required.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>