CA-321983: Handle upgrade case in get_vgpu_compatibility_metadata
VMs that were migrated in from older versions of the software may not have the
VGPU UUID in their NVML record. In fact, xenopsd sets this NVML field to a
random UUID in case the VGPU UUID is unavailable. To handle this case, we call
gpumon with the special empty-string value for the UUID, which causes it to
ignore it.

Note that we still expect exactly one piece of metadata back from gpumon,
because this upgrade case is relevant for single-vGPU VMs only.

This commit also removes an obsolete comment, and avoids a DB call.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>