CA-322655: only replace reboots with shutdown when throttling a VM that crashes too often
If a VM crashes too often we do not want to endlessly restart it, and
have to give up at some point.
However if the VM is set to actions-after-crash=preserve or
coredump-and-destroy then we should be fine executing those actions,
since those are the ones needed to actually debug why a VM is crashing.

Here are all the crash behaviours from xapi:
```
    let on_crash_behaviour = function
      | `preserve -> [ Vm.Pause ]
      | `coredump_and_restart -> [ Vm.Coredump; Vm.Start ]
      | `coredump_and_destroy -> [ Vm.Coredump; Vm.Shutdown ]
      | `restart
      | `rename_restart -> [ Vm.Start ]
      | `destroy -> [ Vm.Shutdown ] in
```

So just replace Vm.Start with Vm.Shutdown to throttle but keep all other
actions intact.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>