CA-322682: Remove PVS Proxy from a VM when it is templated
From CA-322682, In a PVS Environment, if a VM has a proxy enabled for it, and the vm is then converted into a template, the proxy isn't removed. This then means that the template has an existing vif, and as such no further streamed pvs vms can be created from the template as the device already exists.

Signed-off-by: Richard Davies <richard.davies@citrix.com>