CA-322749: Make toolstack work with old Nvidia host drivers
This commit make xenopsd work with old Nvidia drivers as followings,
When VM start, xenopsd will check whether the config_file element exists
in the vm metadata, if yes, then xenopsd think it starts the VM in the legnecy
mode and prepare the config_file arguments for demu, otherwise, the
type_id for the vGPU is used. Xapi is responsible to check whether the
installed host driver support multiple vGPU and set the metadata.

Signed-off-by: Lin Liu <lin.liu@citrix.com>