CA-319960: Remove user KRBTGT from cache before checking
PBIS will return success when it queries user KRBTGT even if the domain is
offline but the user KRBTGT is in the cache. This makes the XAPI fail to
check the availability of domain by this appoach.

This commit adds logic to remove the user KRBTGT from cache before XAPI
checks the availabililty of domain.

Signed-off-by: Ming Lu <ming.lu@citrix.com>