CA-316165: uplift Thread.delay thresholds to make relative timing more reliable
Some of these time values are set too aggressive (small). So in an extremely
stressed environment,  the deviations caused by the heavy loads could become
comparable/noticeable, and in some cases this lead to the break through of
predefined timeboxes or the change of relative order between steps.

We reviewed these cases one by one. For those in need, we uplift the relevant
delay thresholds, to make the steps higher, so that such issues are less likely.
Some cases already have relatively high thresholds set, which remain untouched.

Signed-off-by: Zheng Li <zheng.li3@citrix.com>