CA-316165: Convert some Thread.delay to Delay.wait
Following suggestions on the ticket, here we convert some unconditional wait
(Thread.delay) to conditional wait (Delay.wait).

The pattern is, in these tests we expect to observe something to happen, but it
has to be within a reasonable time frame. From testing point of view, we can
not assume a function will always return (especially if it was implemented
wrong) and we certainly don't want to wait indefinitely.

Previously with Thread.delay, we unconditionally wait for a timeout, then check
the observed value. Not pretty, but it doesn't require any kind of
collaboration from the observed party. The problem is, if we set the timeout
tight, it might time out if the system is extremely stressed. But if we set the
timeout too relax, we end up wasting time every single run.

Now With Delay.wait, we wait conditionally on both a timeout and a
signal/condition. In these case, we can now set the timeout generously (say,
times higher than the usual value, as a real hard timeout), but knowing that in
most cases the wait will end much earlier as usual (upon some
signal/condition) instead of breaking out of the timeout.

Signed-off-by: Zheng Li <zheng.li3@citrix.com>