REQ-627 Detect NVIDIA SRIOV GPUs
We need to know which PCI devices represent NVIDIA SRIOV GPUS. This
commit extends the datatype in xapi_vgpu_type to represent it but does
not yet update the code that parses config files to detect this. All
cards so far assumed to not be SRIOV capable. This will have to be
updated in a later commit.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>