CA-322498: Ensure that QEMU receives the max-memory value used for domain build
QEMU currently gets build_info.Domain.memory_max, which is equal to static-max.
However, the value that is given to xenguest as `-mem_max_mib` to build the
domain is equal to static-max minus video memory. So QEMU is getting passed too
large a number, which can lead to it crashing due to DMA requests for memory
locations that the guest does not actually own.

This patch ensures that QEMU is given the same memory size as xenguest. Note
that there was no problem on qemu-trad, even though it received the same memory
parameter: qemu-trad did not actually use it. We also don't have to worry about
live-migration across versions, where the destination's QEMU is given a lower
memory size, owing to the way this parameter is used within QEMU (according to
the QEMU experts).

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>