CA-322045: Enable fastpath for xcp-rrdd http server
Discovered while investigating CA-322045.

XAPI already enables this by default, so it should be safe to enable it
for xcp-rrdd too.

By default without fastpath Http_svr reads the http headers
byte-by-byte.
With fastpath it reads the minimum HTTP header request size, and then 4
byte at a time until end of headers (it still needs to stop exactly
where the headers finish so that file descriptor passing would work).

It might be possible to improve this further, since AFAICT xcp-rrdd does
not use file descriptor passing on its main socket, but perhaps
we could do that by switching to message-switch or cohttp in the future.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>