CP-30618: Prevent calls to xenopsd with incompatible xen/libxc
This is done by raising a new API error, `XEN_INCOMPATIBLE`, just before the
call to xenopsd, from inside the rpc function in `Xapi_xenops_queue` (this
catches all cases).

This will very quickly kill the startup sequence due to the several xenopsd
calls being done in it, especially in dbsync, so the exceptions are handled
there.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>