CA-322045: tell XAPI to shut down only once
On shutdown XAPI asks xcp-rrdd to backup the RRDs, which can be slow if you have a lot of VMs, SRs, etc.
The shutdown script then asked xapi to shutdown again a second later, which asked xcp-rrdd to save a backup again,
and so on, until finally it gave up after a minute, finally allowing xcp-rrdd to catch up.

If 'xe host-shutdown-agent' succeeds then do not try to repeatedly tell xapi to shut down again,
it heard us the first time. If it fails then do try to tell it to shutdown again.
Either way keep looping until we detect that xapi is down.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>