CA-322045: make light_fuse idempotent
light_fuse* can trigger reboots and saving of RRDs and the database.
It is enough to do this once, even if we are asked to do it multiple
times (in fact it is counter-productive to do it multiple times because
        we would queueu up N RRD backups when we only wanted one).

Rebooting the system would eventually tell xapi to shutdown and trigger
the other fuse, hence the need for different fuse types: Exit when it is
for xapi itself, and Reboot when it is about the system.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>