sr_health_check: Actually make the thread start instead of dead code
OCaml 4.08.1+rc1 found this:
```
File "ocaml/xapi/xapi_sr_operations.ml", line 236, characters 18-705:
236 | ..................Thread.create (fun () ->
237 |               let rec loop () =
238 |                 Thread.delay 30.;
239 |                 let info = C.SR.stat dbg (Storage_interface.Sr.of_string (Db.SR.get_uuid ~__context ~self)) in
240 |                 if not (Db.Task.get_status ~__context ~self:task = `cancelling) &&
...
247 |                 end
248 |               in
249 |               loop ()
250 |             )
Error (warning 5): this function application is partial,
```

And it is indeed correct, this is dead code, we never actually created
the thread, because the `_` was of type `unit -> Thread.t`.
Be explicit about the type of the ignored argument and fix it.

None of the SMAPIv3 SRs that we have actually implement a recovering
health check status, but when they do it would be confusing to have this
code that looks sort of OK but be completely broken.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>