CP-28375: Implement soft reset handler for guest kdump support
Shutdown code 5 means that a domain has kdump kernel loaded and is ready
for a jump to it. This is possible if either the domain is crashed
and we need to load a crash kernel or kexec has been called as part of
e.g. a fast reboot. Code 5 needs to be handled by calling a special
operation called "soft reset" for the domain. This will ensure the domain
is ready for the next enlightned Linux kernel to go through its boot
sequence. The side effect of this operation means xenstore and console
event channels are closed for guest domain and need to be recreated.
The rest of the domain state stays untouched meaning kdump boot process
will follow the same path as bugcheck kernel of Windows guests.

For now only support 1 soft reset per domain lifecycle which corresponds to
crash dump usecase. To support indefinite kexec reboots more garbage
collection needs to be implemented.

Additinally, enable new platform feature - xs_reset_watches. That will
tell the guest it can issue a xenstored command to reset watches for
itself on kexec reboot.

Signed-off-by: Igor Druzhinin <igor.druzhinin@citrix.com>