CA-316165: make test_network_event_loop more deterministic
Actually wait for the update_firewall function to run and raise the
exception.

Otherwise there was a race condition here:
```
 Thread.delay delay;
 fail_firewall_update := false;
 assert_not_called "The event loop should wait for the given delay in case of failures" ();
```
If the network event loop didn't wake up during `delay`, but woke up
after `fail_firewall_update` got set to false then it wouldn't raise the
exception, execute and fail the assertion.
Just to fix the race condition it should've been enough to move the
fail_firewall_update after the assert (and similarly for the 2nd
        instance of this).

However to make it more deterministic do not rely on delays at all: add
a 2nd Threadext.Delay condition variable and wait for the expected event
to occur instead (raise exception), before moving on to the 2nd part,
where we no longer expect the exception to be raised.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>