Simplify by removing simplify
The `simplified` parameter of xenopsd backends was invented specifically for
the xenlight backend. When a backend was `simplified`, this meant that certain
sub-operations could be skipped. E.g. a `VM_start` did not need to include
device plug sub-operations, because these were implicitly done by the
`VM_build` sub-operation.

Now that the xenlight backend has been removed, all the remaining backends are
non-simplified, so we can remove this concept and associated code altogether.

This patch takes the opportunity to make the `atomics_of_operation` function a
little easier to read (hopefully), and removes an out-of-date comment about
QEMU.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>