CA-323893: Catch exceptions in Device.PV_Vnc.is_cmdline_valid
In rare cases, an exception raised from this function during a VM reboot can
cause a call to `VM.stat` to fail. If this happens, then xapi may conclude that
the VM is halted, while it is actually rebooting. This is a problem, because
the VM will be marked as halted rather than running after the reboot.

There is a mechanism in xenopsd that always marks a VM as running, in replies
to `VM.stat`, while it is being rebooted, even when the VM is actually down at
the time. In this case it did not, due to the exception, and xapi concluded "VM
state missing" and reset the power-state.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>