REQ-627 CP-32057 narrow definition of "related" PCIs
PCI devices are considered to be related if they only differ in their
function. Related PCI devices are passed to a guest together.
This definition is insufficient when dealing with virtual functions,
which must be excluded.

This commit narrows the definition of "related" accordingly: virtual PCI
devices are excluded. In addition, the commit tries to make the
definition more explicit by defining predicates.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>