CA-325319 Fix host-display script console handling
There were two distinct issues with the host-display script. Firstly, it always
assumed com1 was the serial port in use, when in some cases it may be e.g. com2.
Secondly if you go through a disable/enable cycle on a normal (VGA) setup, the
result would be the loss of `console=hvc0` from the dom0 command line, because
the `set-dom0` command will *replace* all existing entries, not just append one
to the line.

These changes resolve these issues.

Signed-off-by: Alex Brett <alex.brett@citrix.com>