CA-325582: remove SR RRDs from memory after archiving
This gets run on PBD.unplug: if we keep the RRD in memory then
the memory usage of xcp-rrd will slowly grow after a lot of
SR create/PBD unplug cycles.

Remove the RRD immediately even if the archival itself might fail,
because XAPI would ignore archive failures anyway, and we'd end up with
a space leak if we don't remove it.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>