CA-326174: fix race condition between SR.scan and VDI.forget
VDI.forget was not marking the SR with an in-progress operation so SR.scan could've run while VDI.forget was in progress.
After a VDI.forget the VDI could disappear, causing SR.scan to report a failure.

By default only the plug operation is marked as able to execute concurrently, add VDI.forget to that list:
as long as all we do is run VDI.forget in parallel its fine, but as soon as we try to combine it with
something like SR.scan the SR.scan should get OTHER_OPERATION_IN_PROGRESS error.
VDI.forget should be safe to parallelise because it is just a DB operation, and it was parallelized before.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>