REQ-627 CA-325131 add vGPU.PCI to vgpu map for xenopsd
When Xapi calls VM.migrate in xenopsd by RPC the call includes a map
that handles assigning a new PCI address for the PGPU. We need a similar
mechanism for the PCI VF that is part of a vGPU. Otherwise changes in
this field would not reach Xenopsd.

This commits extends the mapping by including the vGPU.PCI value if it
is not null. It does so by using a special key in the map.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>