CA-326244: Store debug levels in Set ref and do not take lock during logging
Since now the log levels are stored in a ref to an immutable set,
it is safe to read it without taking a lock.

Previously we could deadlock in this module if we try to log from a finalizer, exception hook, etc.
that got invoked while we were already holding the loglevel_m mutex.
This could happen in libraries that use the Logs library, because now
logs from the Log library go through this Debug module.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>