Remove spammy log lines
It appears to be very common for devices in dom0 to not have a driver link
(/sys/bus/pci/devices/XX/driver), i.e. to not be bound to a driver. This
currently leads to lots of alarming lines in the logs. However, it is nothing
to worry about, and not worth logging.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>