CA-326241 assign localhost_ref earlier
When xapi repeatedly fails to start it was not
clearing failed dbsync tasks from the database.
Assigning the resident_on field fixes this
for a pool scenario. The easiest way to do this
is to assign the localhost_ref earlier, such that
`exec_with_new_task` always assigns an actual
host ref rather than a null ref.

Signed-off-by: lippirk <ben.anson@citrix.com>