REQ-627 handle multiple SR-IOV vGPUs
update_pci() assumed so far only one SR-IOV vGPU per VM, which is wrong.
This commit simplifies the code and finds the vGPU belonging to a PCI
device, if such a vGPU exists.

The old code included an optimization that checked the PCI class of the
PCI device first in order to avoid database lookups. However, this can
be traded for looking up vGPUs directly.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>