CA-326241 Set resident_on manually for first task
Attempt two to fix Orphaned dbsync tasks not
being cleaned up. The first 'fix' was breaking
and was reverted in
e0aa8df8fce51be8543c77940b5f706a4712f91a

Server_helpers.exec_with_new_task uses the
global localhost_ref to set the resident on field
for the tasks it creates. However this global
variable is actually assigned during the first
task, so we must set it manually in this case.

Signed-off-by: lippirk <ben.anson@citrix.com>