REQ-627 - act on PCI events during migration
This commit re-writes the update_pci() function that acts on PCI events
sent from xenopsd to xapi. Previously all such events were ignored
during migration but we need to act on them in order to update a VM's
attached PCIs.

The code updating vGPUs is still guarded such that theu are not updated
during migration.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>