REQ-627 CA-328075 after migration, remove stale PCI
This commit adds code to remove stale entries in a VM's list of attached
PCI devices. The code assumes that after a migration the only PCI
devices attached to a VM can be vGPUs because PCI devices can't be
migrated so far.

A conceptually a cleaner solution would use the event mechanism from
xenopsd. However, this would be much more complex and difficult to get
right: during migration xenopsd maintains two temporary identities
(source, destination) of the VM during migration. This makes it
difficult to send (PCI) events under an identity that xapi knows about.
In order to avoid confusion, xapi currently ignores such events during
migration. It would at least for PCI devices need to accept and act on
them. The identity of a domain/VM in xenopsd is quite fundamental and
for this reason difficult to change or to maintain an additional ID for
sending events to xapi.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>