CP-32124: Remove 'LEGACY_CLIENT_ACCEPT' for stunnel
There is a patch against stunnel-4.56 is to accept legacy TLS clients.
The legacy TLS clients are the ones which don't support re-designed TLS
renegotiation caused by CVE-2009-3555.

In stunnel-4.56, XenServer used this patch to support legacy TLS
clients (SSL-662). While upgrading to stunnel-5.55, it is determined
that these legacy TLS clients will not be supported anymore in master.

Therefore, the patch is removed. Accordingly, the usage of this patch,
setting stunnel's option `LEGACY_CLIENT_ACCEPT` is removed also.

Signed-off-by: Ming Lu <ming.lu@citrix.com>