CP-32446: Allow migrated guests to continue using HLE and RTM
In Host.stat compute a mask that we would use for migration CPUID safety checks.
If the TSX_CTRL bit (7) in MSR_ARCH_CAPS is set then it is safe to use
the HLE and RTM CPUID features (i.e. TSX), regardless of the state of this bit
in the host CPUID, so ignore these for migration checks.

This will keep guests working correctly even if they have seen and used
TSX instructions, since the patch in Xen/microcode will still behave
according to the TSX spec (albeit slowly since it will always abort,
and cause guest to fallback to non-TSX codepaths).

After a guest shutdown and startup we'll want to hide the CPUID feature from the guests so they do not attempt
to use TSX anymore and regain some speed.
If TSX is critical to a guest's operation then consult the Xen documentation
on how to enable it and the security tradeoffs.

If the hypercall is not available (e.g. unpatched Xen) then do nothing.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>