CA-325940 bind PCI before calling sriov-manage
Xapi calls Nvidia's sriov-manage on a PCI to activate PCI VFs. If
previously the PCI device was used as pass-through, it is still bound to
the pciback driver. It needs to be unbound and bound to the "nvidia"
driver first.

Unfortunately this duplicates functionality in xenopsd. This is hard to
avoid because xapi needs to activate PCI devices and create entries for
them in the xapi database.

When trying to allocate a PCI VF for an NVidia SRIOV GPU, xapi looks
into its database. The database may contain an entry for a VF that
actually does not exist. This happens, when previously the corresponding
PCI PF was used in pass-through mode. Therefore, xapi needs to call
nvidia_vf_setup in every case. The setup function (which runs on the
relevant host) checks for the presence of VFs before activating them.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>