CP-32437: call assert_can_boot_here with appropriate do_cpuid_check
* `get_possible_hosts_for_vm`: this is used by Xapi_bond and a public API
through `get_possible_hosts`.
None of them talk about migration, so do not perform CPUID checks.
* `retrieve_wlb_recommendations`: was not doing a CPUID check before,
    but I think that is a bug, because `choose_host_for_vm_no_wlb`
    called `vm_can_run_on_host` which did a CPUID check. So do a CPUID
    check here always too.
* `VM.assert_can_boot_here`: this is public API that we modify: always
do the check
* `compute_evacuation_plan_no_wlb`: evacuation implies migration, so
CPUID checks are needed (I think this wasn't correctly handled before)
* `host_of_non_agile_vm`: this is HA related, and HA recovery implies
restarting a VM, not migration: no CPUID checks.
* `pool_migrate`: do the check, unless `force` was used
* `reserve_memory_for_vm`: this can be called from many different
places, some of which can force resuming. This gets called only once
we've decided which host to use, so there is no point in doing the check
here: the CPUID check was already done before, or it'll be done after by
the implementation of the API call.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>