CA-329442: Avoid recreating bridges unless absolutely necessary
When asked to create a bridge, when need to destroy an existing bridge with the
same name _only_ when a the existing bridge is a "real" (non-VLAN) bridge, and
the new one is a "fake" VLAN bridge. Otherwise, the function
`OVS.create_bridge`, which must be idempotent, should leave the bridge alone to
as to not disturb the system.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>