CP-32446: drop support for RPU from pre-Dundee hosts in CPU leveling
CPU leveling code ignored hosts that had missing feature keys
(for compatibility with pre-Dundee that didn't have them).

However we are making changes to these fields now, and this code would
silently ignore any exceptions and CPUID safety checks, which could
cause VMs to crash during an RPU should we have a bug in the code that
handles the feature fields.

Make the code more robust, so that it raises an exception if we fail to
fetch the v2 leveling fields.

If the new `features_hvm_host` field is missing the previous commits
ensure that the code falls back to fetching `features_hvm`.
However if both are missing that will now be an error (or if there is
something wrong with the default fallback path).

Signed-off-by: Edwin Török <edvin.torok@citrix.com>