CP-32446: Support extra CPUID features for migration
Usually a host either supports a CPUID feature or it doesn't.
If it doesn't then it is not safe to migrate a guest that has already
seen the CPUID feature, because applications that use that feature would
crash with an illegal instruction.

Recent hotfixes introduce a third state for a CPUID feature:
* available but slow (due to security mitigations)

Since CPUID features are still expressed through bits they can only
store a binary state, so the way this is handled is:
* CPU exposes new Model Specific Register (MSR), and Xen provides to query
this through host policy hypercall
* When a Xen livepatch is applied the CPUID doesn't change, but
the CPU feature is made safe but slow
* When the host is rebooted Xen hides the CPUID feature that is now
slow from the host, which will cause any newly booted VMs to not see it due to pool
leveling

Previously migration of VMs would be rejected in such a state, because
they've already seen the CPUID feature on boot, but after migration the
host no longer exposes the CPUID feature.
However the MSR tells us that the CPUID feature still works, albeit
slowly, so it is safe to perform such migrations.

Change `features_{hvm,pv}` to report the migration-safe CPUID flags,
and introduce a new internal field `features_{pv,hvm}_host` that reports
the true host flags
(to be used by pool leveling and initial VM boot time CPUID flags).
When MSR leveling is done this internal flag will be replaced by another
one.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>