CA-330830: ignore errors during VM.pause on hard shutdown
In particular Domain_not_built.
This fixes a regression introduced by a7563620e89a891c5a454c1d932ac3f12d0c4c3a.

VM_check_state looks at the domain and what actions should be performed.
On an incomplete migration you might get a VM_shutdown as required
action. However VM_shutdown calls VM_pause, which now fails because the
domain was not built yet.
We should let VM_shutdown proceed in this case to finish cleaning up:
if the domain is not built then there is nothing to pause, we ensured
that it is not running.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>