CA-328093: Correctly report PV-driver status for UEFI-booted VMs
Currently, xenopsd concludes that a VM has PV drivers installed if it has seen
on of the VM's devices (VIFs or VBDs) reach state 4 ("connected"). This is
recorded in the VM's persistent state in field `pv_drivers_detected`, which is
reflected in xapi's `VM.PV_drivers_detected` field, and is not changed until
the VM reboots.

When a VM boots with UEFI, it briefly gets a small VBD attached to it early in
the boot sequence, which has a PV block driver installed and loaded at boot.
Xenopsd will notice this and set `pv_drivers_detected` to `true`, and will not
reassess this state. This is the case, even if the VM's main boot disk does not
have PV drivers installed, and as a result the PV driver status is misreported.

To solve this, xenopsd now considers whether `pv_drivers_detected` must be
reset to `false` if it sees a device reach a state other than 4. Because there
are generally multiple devices, the state of other devices must be taken into
account as well.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>