CA-330961 Clean the yum cache before prechecking/applying an update
It is possible to have two update ISOs which have the same update name, but
contain different versions of the update. The most obvious example is a driver
disk, where we might have different versions depending on the platform version
in use.

There is a potential issue in that yum will cache the repository metadata based
on the name of the repository (which in our case is the name of the update).
If the repository 'on disk' has an older timestamp than the cached data, then
yum will ignore the 'on disk' data, and use the cached information.

In the scenario where a user has erroneously attempted to apply the update from
the wrong version, we can end up in a situation where despite attempting to
apply the correct update, the metadata has been cached from the incorrect one,
and thus yum will refuse to install it due to a dependency failure (this is what
happened in XSI-539).

To avoid this, clear the yum cache before we attempt a precheck or update apply.

Signed-off-by: Alex Brett <alex.brett@citrix.com>