CA-324461 don't always make resync_host api call
Previously resyncing a host (within context of
clustering) would always make an api call. This
is not what we want when the resync operation
is running on the host we want to resync.
As a solution we make the resync_host caller
decide whether to make an api call or not. This
has been wrapped in a helper.

Signed-off-by: lippirk <ben.anson@citrix.com>