CA-330979: set current_domain_type for slaves
ff50be0b3b20102ca49cd252d32bdaea89619e06 made it an error to have an
unspecified current domain type at VBD plug time.

Dom0 of slave hosts always had an unspecified current_domain_type,
a bug caused by the pool join code.

When a slave joins a pool we import all VMs, and the import code deals
only with halted VMs (it sets current_domain_type to `unspecified`).
However Dom0 is special: it is the only VM allowed to be running when
joining a pool, so we must fix up the current_domain_type.
Do this in the function that updates other fields of Dom0's metric
record on startup to ensure it is always correct.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>