CA-325582: Remove SR DSs from memory on SR.forget and destroy
Make sure data sources for an SR are not kept in memory in xcp-rrdd
when the SR is destroyed or forgotten.

sr-destroy is separated into 3 stages: the first is
executed by the SR master, which physically destroys the contents of the
SR, the second is executed by all the hosts that have had a PBD,
which cleans the metrics about them. And the third one is the database
cleanup, executed by the pool master.

We need to change `assert_all_pbds_unplugged` to tell us whether we have
any PBDS or not, since if we don't have any PBDs, then there are no SR
metrics to remove.

sr-forget has been separated into 2 steps, which correspond to the last
2 of sr-destroy.

We also make `consider_unplugged_pbds` a non-optional parameter to
`choose_pbd_for_sr`, as it really ought not to be (it only has a few
clients).

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>
Signed-off-by: lippirk <ben.anson@citrix.com>
(cherry picked from commit deb0b6c13f4692cb76adce08c662770cb66f1c8e)